requirements.txt
fastapi
uvicorn
python-dotenv
qdrant-client
google-generativeai
pypdf
python-docx
